Add test for sensor subscriptions
This will break if anything in the sensor names, for example, changes.
Otherwise, the test that no messages are received during the blackout
is not conclusive and may pass by accident.

Also, including tf_sensors.yaml will over-write the sensors from sample_user_config,
so the logical camera is no longer being tested.